# otelslog: Remove the pooled buffer