Fixes https://github.com/open-telemetry/opentelemetry-go-contrib/issues/5879

The `sync.Pool` in `otelslog` gives no performance improvements.

Moreover from https://pkg.go.dev/go.opentelemetry.io/otel/log#Logger `Emit` docs:

> The record may be held by the implementation. Callers should not mutate the record after passed

This would mean that we cannot pool the attributes as it can be held by the implementation (e.g. batch processor).

```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/contrib/bridges/otelslog
cpu: Intel(R) Core(TM) i9-10885H CPU @ 2.40GHz
                                           │   old.txt    │               new.txt                │
                                           │    sec/op    │    sec/op     vs base                │
Handler/Handle-16                            84.75n ± 23%   80.11n ± 55%        ~ (p=0.912 n=10)
Handler/WithAttrs/5-16                       420.8n ±  8%   449.9n ± 18%        ~ (p=0.684 n=10)
Handler/WithAttrs/10-16                      667.8n ± 13%   681.2n ± 18%        ~ (p=0.836 n=10)
Handler/WithGroup-16                         134.4n ± 18%   102.1n ±  3%  -24.03% (p=0.000 n=10)
Handler/WithGroup.WithAttrs/5-16             570.0n ±  7%   440.6n ± 29%  -22.69% (p=0.002 n=10)
Handler/WithGroup.WithAttrs/10-16            885.7n ± 22%   646.5n ±  7%  -27.01% (p=0.000 n=10)
Handler/(WithGroup.WithAttrs).Handle/5-16    168.0n ± 24%   126.3n ± 17%  -24.82% (p=0.001 n=10)
Handler/(WithGroup.WithAttrs).Handle/10-16   175.4n ± 33%   141.4n ± 23%        ~ (p=0.052 n=10)
geomean                                      288.2n         244.5n        -15.18%

                                           │   old.txt    │               new.txt               │
                                           │     B/op     │    B/op     vs base                 │
Handler/Handle-16                            0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Handler/WithAttrs/5-16                       264.0 ± 0%     264.0 ± 0%       ~ (p=1.000 n=10) ¹
Handler/WithAttrs/10-16                      472.0 ± 0%     472.0 ± 0%       ~ (p=1.000 n=10) ¹
Handler/WithGroup-16                         64.00 ± 0%     64.00 ± 0%       ~ (p=1.000 n=10) ¹
Handler/WithGroup.WithAttrs/5-16             328.0 ± 0%     328.0 ± 0%       ~ (p=1.000 n=10) ¹
Handler/WithGroup.WithAttrs/10-16            536.0 ± 0%     536.0 ± 0%       ~ (p=1.000 n=10) ¹
Handler/(WithGroup.WithAttrs).Handle/5-16    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Handler/(WithGroup.WithAttrs).Handle/10-16   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                 ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                           │   old.txt    │               new.txt               │
                                           │  allocs/op   │ allocs/op   vs base                 │
Handler/Handle-16                            0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Handler/WithAttrs/5-16                       3.000 ± 0%     3.000 ± 0%       ~ (p=1.000 n=10) ¹
Handler/WithAttrs/10-16                      3.000 ± 0%     3.000 ± 0%       ~ (p=1.000 n=10) ¹
Handler/WithGroup-16                         2.000 ± 0%     2.000 ± 0%       ~ (p=1.000 n=10) ¹
Handler/WithGroup.WithAttrs/5-16             5.000 ± 0%     5.000 ± 0%       ~ (p=1.000 n=10) ¹
Handler/WithGroup.WithAttrs/10-16            5.000 ± 0%     5.000 ± 0%       ~ (p=1.000 n=10) ¹
Handler/(WithGroup.WithAttrs).Handle/5-16    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Handler/(WithGroup.WithAttrs).Handle/10-16   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                 ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```