@pellared 2024-07-09
I am skipping the changelog as there is only some performance boost. Not sure if it is worth mentioning.